\name{reams-package}
\alias{reams-package}
\alias{reams}
\docType{package}
\title{
Resampling-based adaptive model selection}
\description{
Resampling methods for adaptive linear model selection.  These can be thought of as extensions of the Akaike information criterion that account for searching among candidate models.  A number of functions in the package depend crucially on the \pkg{leaps} package, whose authors are gratefully acknowledged.
}
\details{For a complete list of functions type \code{library(help=reams)}.
}
\author{
Philip Reiss <phil.reiss@nyumc.org> and Lei Huang <huangracer@gmail.com>

Maintainer: Tao Zhang <tao-zhang-1@uiowa.edu> }
\references{
}
\keyword{package}
\seealso{
}
\examples{
}
